+++ manifesto/mnfs_parser.py
